pyproject.toml
[tool.poetry]
name = "join-to-it-interview"
version = "1.0.0"
description = ""
authors = ["Nazar Lytvynov <ratdesertwot@gmail.com>"]
readme = "README.md"
packages = [{ include = "src" }]

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.0.2"
djangorestframework = "^3.14.0"
drf-spectacular = "^0.27.1"
django-filter = "^23.5"
python-dateutil = "^2.8.2"


[tool.poetry.group.dev.dependencies]
mypy = "^1.8.0"
isort = "^5.13.2"
django-stubs = "^4.2.7"
djangorestframework-stubs = "^3.14.5"
pytest = "^8.0.1"
pytest-django = "^4.8.0"
factory-boy = "^3.3.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 120
include = '\.pyi?$'
exclude = '''
/(
    pycache
  | migrations
)/
'''

[tool.isort]
profile = "black"

[tool.mypy]
files = "src, tests"
plugins = ["mypy_django_plugin.main", "mypy_drf_plugin.main"]
exclude = 'migrations/'

[tool.django-stubs]
django_settings_module = "src.settings"

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = "tests.py test_*.py *_tests.py"
DJANGO_SETTINGS_MODULE = "src.settings"
django_debug_mode = "keep"
